Rollback the change made in the field 'property_price'.

+++ property/models/Property.py
@@ -14,7 +14,7 @@ class Property(models.Model):
     property_category = models.IntegerField(
         choices=PROPERTY_CATEGORY, default=2)
     property_type = models.IntegerField(choices=PROPERTY_TYPE, default=1)
-    property_price = models.FloatField()
+    property_price = models.IntegerField()
     bedrooms = models.IntegerField()
     bathrooms = models.IntegerField()
     metreage = models.FloatField()
